Add a Reason metadata to the books
I want to fill in the reason of a book that I plan to read, because
sometimes it takes a while before I can actually start reading it. If
the book doesn't make sense at a certain point in time, then I can pause
reading it for even longer or throw away if I don't deem it valuable anymore.